Code improvements for graphing

+++ test/test_graph_calculations.cpp
@@ -64,12 +64,12 @@ TEST_CASE("ABV Vector Creation", "[Graph Data Compilation]") {
 
 TEST_CASE("Count Values in Vector", "[Graph Data Compilation]") {
     std::vector<double> input_values{5.5, 5.5, 6.0, 10.0, 5.5, 4.8, 4.8, 10.0, 10.0, 10.0};
-    std::map<double, int> expected_values = {{4.8, 2},
+    std::map<double, size_t> expected_values = {{4.8, 2},
                                              {5.5, 3},
                                              {6.0, 1},
                                              {10.0, 4}};
 
-    std::map<double, int> output_values = Graphing::count_values_in_vect(input_values);
+    std::map<double, size_t> output_values = Graphing::count_values_in_vect(input_values);
 
     REQUIRE(output_values == expected_values);
 }
